+++ frontend/src/app/(app)/(addEmployee)/people/add-new/layout.tsx
@@ -1,5 +1,6 @@
 import { EmployeeSideStepper } from '@app/components/modules/EmployeeSideStepper';
 import { WorkflowTopbar } from '@app/components/modules/WorkflowTopbar';
+import { AddEmployeeFormProvider } from '@app/components/pages/addEmployee/AddEmployeeFormProvider';
 
 export default function PeopleLayout({ children }: Readonly<{ children: React.ReactNode }>) {
   return (
@@ -9,7 +10,9 @@ export default function PeopleLayout({ children }: Readonly<{ children: React.Re
         <main className="p-8">
           <div className="grid grid-cols-workflow">
             <EmployeeSideStepper />
-            <div className="col-span-1 px-8">{children}</div>
+            <div className="col-span-1 px-8">
+              <AddEmployeeFormProvider>{children}</AddEmployeeFormProvider>
+            </div>
           </div>
         </main>
       </div>
