+++ frontend/src/app/(app)/(addEmployee)/people/add-new/main-ladder/[[...id]]/page.tsx
@@ -1,0 +1,3 @@
+import { MainLadder } from '@app/components/pages/addEmployee/MainLadder';
+
+export default MainLadder;
