+++ frontend/src/app/(app)/(addEmployee)/people/add-new/personal-details/[[...id]]/page.tsx
@@ -1,0 +1,3 @@
+import { PersonalDetails } from '@app/components/pages/addEmployee/PersonalDetails';
+
+export default PersonalDetails;
