--- frontend/src/app/(app)/(peopleFlow)/people/add-new/main-ladder/page.tsx
@@ -1,3 +1,0 @@
-import { MainLadder } from '@app/components/pages/MainLadder';
-
-export default MainLadder;
