--- frontend/src/app/(app)/(peopleFlow)/people/add-new/personal-details/page.tsx
@@ -1,3 +1,0 @@
-import { PersonalDetails } from '@app/components/pages/PersonalDetails/PersonalDetails';
-
-export default PersonalDetails;
