+++ frontend/src/components/pages/addEmployee/MainLadder/MainLadder.utils.tsx
